requirements.txt
streamlit>=1.35.0
requests>=2.31.0
google-generativeai>=0.7.0
resend>=2.0.0
python-dotenv>=1.0.0
